+++ tools-public/toolpad/pages/joyUiNpmKpis/page.yml
@@ -8,7 +8,7 @@ spec:
   queries:
     - name: queryJoyUIDownloads
       query:
-        function: queryJoyUIDownloads
+        function: queryJoyUIMonthlyDownloads.ts#queryJoyUIDownloads
         kind: local
   content:
     - component: Text
